requirements.txt
ollama
fastapi 
uvicorn 
pydantic 
starlette 
rich 
pillow  
python-multipart
chromadb
groq
streamlit  
python-dotenv 
google.generativeai